fix error in main-cs module
this code needed to be touched, because jshint didn’t like the old way it
was written. Unfortunately the "fix" was wrong, of course the check and the
has to be _inside_ of the function. This prevented the plugin from working
in firefox and everything seems to work fine, again.