fix logic which password dialog to open when signing
after a specific key is selected, the user is either in webmail mode
or external editor mode. Signing is nothing to do with decryption settings.
Now the test is correct and a popup is opened for webmail mode and the
password dialog is shown inline when in external editor mode.