make password dialog really modal in external editor
solution: hide modal editor dialog when pwd dialog shown

bootstrap behaves a little strange when two modal dialogs are open at the
same time. This should be refactored, later, but for now it looks as
intended. At least with all the nested iframes, it’s hard to get it right.
We should switch to a more template based approach, in the future.