add grunt-bump to support releases
in order to help during the release process, the grunt-bump can be helpful.
Once configured, it will be possible to bump the version number in all files
with just one command and commit these changes. It's also possible to tag
and push the commits directly, but these are deactivated, since bumping
versions in all files might not be enough.