Move the encryptframe button not beyond the parent
With my webmailer I've some problem at this part.  
The substraction `- toolbarWidth - 20` moves the _eFrame to left position -23.  
I think there is some problem to detect the correct editElementWidth.  
The result is, that I can not see the button. It is under some other element of the webmailer.  
But the frame around the button is from -23 to the right border.  
My patch is a fast workaround. I think a negative left-position makes no sense at all and so you can see at least the button.  
Maybe you'll find the real bug and can fix it better.