tests: keyServer class was replaced by mveloKeyServer module
The tests were not yet adjusted to this.
importing the non existing file led to tests not being run anymore.

keyserver-test itself has seen a bit of rework
to adjust it to testing a module rather than a class.

Also dropped setter for baseURL from mveloKeyServer.
This was only used in the tests.