tests: fix key server test
The tests for the _url function were failing
since it is not exported anymore.

Instead of testing the _url function
we now check that the urls used in the lookup, upload and remove calls
are indeed correct.
We do so by using the window.fetch stubs as spies
and checking the arguments.