autocrypt: Introduce autocryptWrapper
For the lack of a better name...

autocryptWrapper wraps the storage and the callback based calls
to autocrypt-js Autocrypt class in an asynchronous API.

Starting with processing headers and looking up keys for now.

Right now the storage is ephemeral and only works in tests.
We will probably want to reuse mailvelopes storage.
But that is for another step.