one storage entry for ac per keyring
Store all autocrypt data for a given keyring in one mvelo.storage entry.

We used to have different entries for every new email address.

This commit introduces a Store class on top of mvelo.storage.
A store uses a Map internally but serializes it as an object.
It currently has a callback based API so we can use it with Autocryp.
Will default to an async API for use in KeyringAttrMap
and expose the callback API separately.