autocrypt: add header and footer to keys
Autocrypt headers leave out the BEGIN PGP ... and the END PGP ... lines.
They also join all the lines together and leave out the checksum.

We turn the base64 encoded string we get from the headers into a proper armored key.
We can not bring back the checksum but it is not required anyway.