Initial test for the API
API level tests allow testing the wiring in the event handling
inside the controllers and the content scripts.
It ensures that API calls actually accomplish their goals.

Polyfills and Stubs
-------------------

test/polyfills.js now includes a simple implementation of
chrome.runtime.connect and .onConnect.

The local storage is stubbed entirely so we do not have to worry
about replacing it with rewire in some contexts.
We do not recover it after the run because there is no
working local storage in the test environment anyway.

Loading mailvelope-client-api.js
--------------------------------

content-scripts/clientAPI.js will add a script tag to the head
which then tries to load the mailvelope-client-api.js
This used to result in the following warning:
WARN [web-server]: 404: /context.html/client-API/mailvelope-client-api.js

We now have a proxy defined in test/karma.conf.js.
An additional preprocessor line was needed as
`'**/*.js'` will not include `'../src/client-API/**/*.js'`.

Make keyring.init clear internal state
--------------------------------------

We keep the keyringMap and keyringAttr as constants in the keyring closure.
This way they act as singletons that get initialized at load time
and then they keep some state that is also synchronized to the store.

During normal runs this is fine. But in tests the state in keyringMap
and keyringAttr gets out of sync if we clear the locale storage.
This way test state leaks from one test to the next.

The keyring test even made use of this creating a keyring in one test
and then depending on it in another.

This change clears the internal state of the keyring on init.

The dependency between the tests has also been removed
and we now have both tests run based on the same data
created during the initial beforeEach hooks.

Make main.initController not async
----------------------------------

initController only registers callbacks.
There's no need for this to be async.

This also makes running it in tests way easier.