get new tests to run for now
api.init() now needs prefs to be defined with a version.
This should be moved out of the test case before merging.

I also observed:
25 02 2019 11:39:08.572:WARN [web-server]: 404: /context.html/client-API/mailvelope-client-api.js

Changed the url to have a leading / in clientAPI script.
I have no idea what caused the different behaviour. So need to check this during review.