Gmail ExtractFrame: scroll bar hides close button (#849)
* fix(gmail): scroll bar hides close button

In the top right corner of a extract frame we display a close button. In some cases this close button is hidden by vertical scrollbars on the right side. This was not there before and looks like a style change in Gmail.

The fix also tries to mimic the default behaviour of `decryptFrame` where when in popup mode, the overlay stretches to the whole width of the message instead of fixing width and height.

* docs: refine comments

* docs: fix typo in a comment